# Steam Apps: Don't fetch soundtracks as steamapps