@sonic2kk 2023-02-09
Actually just noticed when re-reading after posting with the screenshots, that Proton Next appears in the game list too. I'd like to tackle that in this PR as well :-)