@DavidoTek 2023-02-10
Yeah, makes a lot of sense to not include the soundtracks in the game dialog ;)

Your approach seems good. I also cannot think of another way to get this information, `appinfo.vdf` probably doesn't have any infos on this?
I'm wondering, does the approach have any impact on the start time?

> ProtonUp-Qt does not detect Proton Next as a separate compat tool

Steam own compatibility tools are read in a different way using [`get_steam_ctool_list`](https://github.com/DavidoTek/ProtonUp-Qt/blob/6f42a7d03e69caa540d86576642144e35bd441d8/pupgui2/steamutil.py#L149). It obtains the information whether a Steam app is a compatibility tool or not from the `appinfo.vdf` (see [this](https://github.com/DavidoTek/ProtonUp-Qt/blob/6f42a7d03e69caa540d86576642144e35bd441d8/pupgui2/steamutil.py#L182)). Maybe this contains incorrect information about `Proton Next`.

Thanks!