@sonic2kk 2023-02-10
> `appinfo.vdf` probably doesn't have any infos on this?

From what I could see, it does not unfortunately. I couldn't find anywhere that stored information about app type :frowning: 

>  I'm wondering, does the approach have any impact on the start time?

So when I first thought of this approach I also thought will this impact startup time, since so much work was put in recently in improving that I didn't want to just step all over it.

In my usage I didn't notice a bump, but you put me in mind to actually time it :-) So here we go!

I tested by putting a `time.time()` at the start and end of `get_steam_app_list`, where the additional logic is placed.

#### Main Branch
- First load takes roughly 2.3 seconds.
- Opening games dialog takes roughly 1.35s

<details>
<summary>Full testing times:</summary>
<hr>
First load of ProtonUp-Qt:

- 1st test: `get_steam_app_list took 2.344550371170044 seconds`
- 2nd test: `get_steam_app_list took 2.302269220352173 seconds`
- 3rd test: `get_steam_app_list took 2.3538291454315186 seconds`

Takes roughly 2.3 seconds.

Opening Games List dialog (runs twice so showing both values):
- 1st test:
    - `get_steam_app_list took 1.326843500137329 seconds`
    - `get_steam_app_list took 1.3295726776123047 seconds`
- 2nd test
    - `get_steam_app_list took 1.3618791103363037 seconds`
    - `get_steam_app_list took 1.3580753803253174 seconds`
- 3rd test
    - `get_steam_app_list took 1.3741896152496338 seconds`
    - `get_steam_app_list took 1.3622820377349854 seconds`

Takes roughly 1.35 seconds.
<hr>
</details>

#### This PR
- First load takes approximately 2.4s.
- Opening games dialog takes roughly 1.4s.

<details>
<summary>Full testing times:</summary>
<hr>
First load of ProtonUp-Qt:

- 1st test: `get_steam_app_list took 2.3772754669189453 seconds`
- 2nd test: `get_steam_app_list took 2.395808458328247 seconds`
- 3rd test: `get_steam_app_list took 2.38689923286438 seconds`

Takes roughly 2.4 seconds.

Opening Games List dialog (runs twice so showing both values):
- 1st test:
    - `get_steam_app_list took 1.4157302379608154 seconds`
    - `get_steam_app_list took 1.3954129219055176 seconds`
- 2nd test
    - `get_steam_app_list took 1.3871181011199951 seconds`
    - `get_steam_app_list took 1.3708641529083252 seconds`
- 3rd test
    - `get_steam_app_list took 1.402207374572754 seconds`
    - `get_steam_app_list took 1.3632738590240479 seconds`

Takes roughly 1.4 seconds.
<hr>
</details>

It seems like, looking at the timings here, that on my PC it takes very slightly longer. This probably is not a big deal, and to my eye I don't even notice the bump in load time. However please feel free to test this more thoroughly, or if there's a better way to parse the info from the file that could reduce the load times I would be happy to improve it :-)

> Maybe this contains incorrect information about Proton Next

You may be onto something here. Proton Next is a weird case. Its folder structure is slightly different (it's missing the unpacked `dist` folder and just has a `tar`), and as well as this, currently its `VERSION` file is identical to that of the current stable Proton 7.0. It could indeed just be held in an unusual way.