@DavidoTek 2023-02-10
That's great. Thanks for taking the time and doing the testing :)

> However please feel free to test this more thoroughly, or if there's a better way to parse the info from the file that could reduce the load times I would be happy to improve it :-)

I did a quick test and parsing a single appmanifest.acf file took less than 1ms on average (actually, parsing a single file 2000x took ~0.2ms per run. Not sure if there's any caching involved in parsing a single file).
I think that's fine and maybe your approach will come in handy in the future if we need any other information from that file.

> You may be onto something here. Proton Next is a weird case.

That's interesting. I'm wondering if its releases are also handled differently, like if it's just a link to the latest not yet released version...